--- front/src/pages/auth/signup/SignUpPage.tsx
@@ -1,74 +1,0 @@
-import React, { useState } from "react";
-import signUp from "@/features/auth/signUp";
-import { useRouter } from "next/router";
-
-const SignUpPage = () => {
-  const [name, setName] = useState("");
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [passwordConfirmation, setPasswordConfirmation] = useState("");
-  const router = useRouter();
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    if (password !== passwordConfirmation) {
-      alert("Passwords do not match");
-      return;
-    }
-    try {
-      await signUp(name, email, password, passwordConfirmation);
-      alert("Sign up successful");
-
-      window.dispatchEvent(new Event("storage"));
-      router.push('/')
-    } catch (error) {
-      alert("Sign up failed");
-    }
-  };
-
-  return (
-    <div>
-      <form onSubmit={handleSubmit}>
-        <div>
-          <label>Name:</label>
-          <input
-            type="text"
-            value={name}
-            onChange={(e) => setName(e.target.value)}
-            required
-          />
-        </div>
-        <div>
-          <label>email:</label>
-          <input
-            type="email"
-            value={email}
-            onChange={(e) => setEmail(e.target.value)}
-            required
-          />
-        </div>
-        <div>
-          <label>password:</label>
-          <input
-            type="password"
-            value={password}
-            onChange={(e) => setPassword(e.target.value)}
-            required
-          />
-        </div>
-        <div>
-          <label>Confirmation password:</label>
-          <input
-            type="password"
-            value={passwordConfirmation}
-            onChange={(e) => setPasswordConfirmation(e.target.value)}
-            required
-          />
-        </div>
-        <button type="submit">Sign Up</button>
-      </form>
-    </div>
-  );
-};
-
-export default SignUpPage;
